Create OEP-7 draft to replace OEP-1, to support Product proposals
The purpose of this OEP is to add support for proposing changes to the
functionality of the Open edX platform to the OEP process. Rather than
using multiple processes, I believe the OEP process can be easily
adapted to this case, particularly with some light support for less technical
forms of discussion.

Changelog:

* New "Product" proposal type
* Remove references to arch@ email address
* Created "Inital Submission" section
* Increased scope of Arbiter role to include helping with GitHub and
  other technical mechanics as needed
* Added support for Google Docs and other external forums for discussion
  of the proposal
* Added "References" field to the table as a place to put links to
  external discussions