Improve guidance on when to start an OEP
I've separated out the steps for brainstorming an OEP & getting feedback from the actual process of submission.  The main reason for this PR is that I'd like clearer guidance as to what kind of suggestions make for good OEPs.  I've added a section for examples but cannot populate it myself, as I do not know what makes for a good or bad OEPs.