doc: OEP-37: Dev Data
This is an approach to be able to generate repeatable test data
for developers to have at hand in their development environment.

We've started working on the implementation as if this is accepted in as
a provisional OEP so we should reflect that and land it.  Further
improvements can still be made as follow on PRs.