Add a horror tale when deadcode is not removed
In the context of dead code, feature toggles could hide old code that if active could carry on significant damage. Here's a reference to an U.S. Securities and Exchange Commission act where they order the cease and desist to a company that lost more than 400 million in financial transactions due to a "repurposed a flag".