Provide guidance for archiving forks multiple times.
There are a lot of options for this but choosing the simplest for now
since this is a pretty corner case and it's unclear how often this will
happen let alone how often it will bite us to not have a more complex
process here.

Currently there are only 5 forks under edx-unsupported that could have
collisions, 4 of which we are unlikely to fork again.