docs: update OEP-17 on feature toggles
These updates are meant to bring OEP-17 more up to date
with the toggle framework advancements to date. It adjusts
some of the details, without really changing the high-level
requirements that were originally decided in the OEP.

* Renamed "Graceful Degradation" to "Circuit Breaker", which was
decided to be a more standard name for this use case.
* Added reference to the new edx-toggles library and its
discoverability and reporting features, including its how-to
documents.
* Updated our use of Django Settings for feature toggles, now that
we can more easily deploy configuration changes.
* Removed some of the links to specific classes in favor of how-tos
which are more likely to be kept up to date.

ARCHBOM-1746