docs: add warning for external ids
Adds a warning and the start of some clarity around cases
where we do not recommend using the internal LMS user_id.

A more formal OEP update should be done in the future,
but this at least brings awareness to the new ADR and
functionality written since the OEP was first written.