fix: Updates to OEP-21: Deprecaton (#207)
We have found that the usage of Removal dates is
somewhat confusing and just duplicates the current functionality
of Accepted dates. We are removing them to help clarify the process.

We would also like to include links to Discourse threads in DEPR
tickets, because it makes it easier to cross reference the discussions.