feat: In OEP-18 recommend putting constraints in `base.in` in libraries
Constraints should be expressed in the base.in file whenever it's
important that a relying package also use that constraint. Otherwise,
there's no way to ensure that an IDA's requirements files are compiled
in a way that respects the known incompatibilities in a library.

Also add some more encouragement to impose minimum version constraints in
libraries when developers notice newly added features are being used.

This change is inspired by an observation of an incident where
edx-proctoring was upgraded to compatibility with PyJWT 2.0.0's breaking
changes but edx-platform was limited to PyJWT<2.0.0, leading to errors in
production. This situation could have been prevented only if:

1. A minimum version constraint on PyJWT was added in edx-proctoring, and
2. The constraint was added in base.in rather than constraints.txt.

Both of these were previously in contravention of this OEP.