docs: Add an alternative link to a publication.
The paper for `What makes a great maintainer of Open Source Projects?`
is behind a pay wall on the IEEE website but it is also available wihout
that restriction on the author's website.  So I added an alternative
link so people can get to the content.

I left the original link because the IEEE publication speaks to the fact
that it is a peer reviewed paper which lends to how much weight the
content carries.