docs: tighten OEP-18 rules on URLs for Python dependencies
Nine different URL-specified Python dependencies had
accumulated in edx-platform, noticably increasing the amount
time it took to install edx-platform requirements.

After many developer hours of work, we are nearly finished
converting them to PyPI-hosted depencencies:
https://github.com/openedx/wg-developer-experience/issues/153

We make these OEP updates in order to clarify that URL
dependencies should not be added except in the two specified
circumstances and that the follow-up work of moving them
to PyPI must be recorded.