docs: adding an ADR to OEP-65 about a unified platform repository (#598)
* docs: adding an ADR to OEP-65 about a unified platform repository

This ADR describes a decision to create a unified platform library, combining frontend-build, frontend-platform, frontend-plugin-framework, frontend-component-header, frontend-component-footer, and eslint-config.  It will also include the ‘shell’ application for the new architecture proposed in OEP-65.

* docs: fixing some RST styling issues and removing link

Link was to a second ADR that hasn’t been written yet.

* docs: misc formatting, adding a change history

* docs: adding a reference to the ADR from OEP-65

* docs: adding a toctree for decisions to OEP-65

* docs: addressing review feedback

* docs: splitting up the “Significant API Changes” section by library

Thanks to @brian-smith-tcril for the suggestion.