docs: Relax CC requirement for maintaining.
This can happpen when there is a team that maintains a repo and the team
is adding a new member.  That new member will not have gone through the
CC process yet but will be listed as a maintainer by virtue of being on
the team in the `catalog-info.yaml` file.  Allow for this to be
possible.  In practice we won't run into a lot of issues with this as
long as at least one person on the maintaining team can merge
maintenance work.