docs: small fixes and improvements for OEP 51, 54, and 67 (#706)
* docs: reword to improve clarity and flow, fix typo

The original wording is fine, but when I first read the document
I felt like this sentence didn't read quite as well as others.

I've tried to improve on the original sentence to add additional
clarity and emphasis.

* fix: fix broken and outdated links

The hyperlinks for commitlint.yml and commitlint.config.js were being
interpreted as inline/relative links rather than external. I removed
the inline syntax and instead just reference the linked text.

The git hyperlinks were referencing outdated anchors, so I have updated
these to be current.

* docs: remove exception for eslint config for es5

As mentioned in #681, this exception is perhaps too narrowly applicable
to be worth mentioning at this point.

* fix: update unused, outdated, and hard links

- remove unused explicit target links
- update link for eslint config to new location
- update edX Frontend Auth Client links and references to frontend-platform (new location)
- use :doc: roles for links to other Open edX documentation rather than hard links

See #681

* fix: expand contractions to improve clarity for non-native English speakers